remove extranious association registry
